day27/pyproject.toml
[project]
name = "ai-requisite-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
    "tenacity>=9.1.4",
]
